Move token write to it's own file
