+++ functions/src/db/isTokenValidInDb.ts
@@ -1,7 +1,7 @@
-import { JWTData } from '../../types/JWTData'
-import { JWTToken } from '../../types/JWTToken'
-import { db } from '../../firebase/firebase'
-import { UserData } from '../../types/UserData'
+import { JWTData } from '../types/JWTData'
+import { JWTToken } from '../types/JWTToken'
+import { db } from '../firebase/firebase'
+import { UserData } from '../types/UserData'
 
 export const isTokenValidInDb = async (
     jwtData: JWTData,
